Move appcache files to gridcoin directory
