+++ src/Makefile.test.include
@@ -23,7 +23,6 @@ OTHER_TEST_FILES = \
 
 # test_n binary #
 GRIDCOIN_TESTS =\
-  test/appcache_tests.cpp \
   test/checkpoints_tests.cpp \
   test/dos_tests.cpp \
   test/accounting_tests.cpp \
@@ -37,6 +36,7 @@ GRIDCOIN_TESTS =\
   test/fs_tests.cpp \
   test/getarg_tests.cpp \
   test/gridcoin_tests.cpp \
+  test/gridcoin/appcache_tests.cpp \
   test/gridcoin/beacon_tests.cpp \
   test/gridcoin/claim_tests.cpp \
   test/gridcoin/contract_tests.cpp \
