+++ src/gridcoin/researcher.cpp
@@ -1,8 +1,8 @@
-#include "appcache.h"
 #include "backup.h"
 #include "boinc.h"
 #include "contract/message.h"
 #include "init.h"
+#include "gridcoin/appcache.h"
 #include "gridcoin/beacon.h"
 #include "gridcoin/magnitude.h"
 #include "gridcoin/project.h"
